Add specialized masks for mass budget calculations
Add grounded and floating masks that are used for the mass budget
calculation and are different from the masks in cellMask. Currently,
the only difference is that the grounded mask includes floating non-dynamic
cells adjacent to grounded ice. The grounded masks will eventually
also include floating ice over subglacial lakes, but that requires
flood_fill to be moved from li_calving to a shared routine, so that
will come in a future commit. In preparation for that, add domain
variable in calls to li_calculate_mask.