Account for multiple updates of calvingThickness within a timestep
Add a very short subroutine update_calving_budget that is called to
recalculate groundedCalvingThickness and floatingCalvingThickness
after calvingThickness is applied and before masks are updated. Also
treat grounded and floating calving specially within the remove_icebergs
and remove_small_islands routines to deal with the face that
calvingThickness is updated multiple times per timestep. Thus, it
is okay for a cell to have both nonzero groundedCalvingThickness
and floatingCalvingThickness.