Update calving budget incrementally
Use a local variable dCalvingThickness to update the calving budget
each time calving is applied. The variable calvingThickness is now
not used directly to modify thickness, but calculated by summing
dCalvingThickness throughout a time step. This might lead to issues
with halo updates because dCalvingThickness is always an allocatable
array. If so, it will need to either be made a Registry variable or
an MPAS allocatable scratch array.