Add SSP-RK3 time integration
Add support for config_rk_order = 3, using the Strong Stability-
Preserving RK3 method. This follows a slightly different algorithm
from the classic RK2 and RK4 methods, and so requires significantl
increasing the complexity of the RK loop. See eq 2.48 (pg 56) in
Durran (2010): Numerical Methods for Fluid Dynamics for the description
of this method.